Feat : 도수 입력 시 소수점도 입력 가능하게 수정

+++ data/src/main/java/com/teamwiney/data/network/model/response/TastingNoteDetail.kt
@@ -15,7 +15,7 @@ data class TastingNoteDetail(
     val color: String,
     val buyAgain: Boolean,
     val varietal: String,
-    val officialAlcohol: Int?,
+    val officialAlcohol: Double?,
     val price: Int?,
     val smellKeywordList: List<String>,
     val myWineTaste: MyWineTaste,
